requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1
langchain==0.2.17
langchain-chroma==0.1.4
langchain-openai==0.1.25
langchain-community==0.2.19
langchainhub>=0.1.21
chromadb>=0.5.5
faiss-cpu>=1.8.0.post1
py2neo>=2021.2.4
pymupdf>=1.24.10
requests>=2.32.3
